Initial Upload
Mirror project at 1:35 AM Pacific time, 7/10/2020. Users sync properly cross-platform. Environment will do for now. Ball (or anything else) does not sync properly because of server authority issues. The solution to this is somewhere in the [command] function, which allows actions to be run on the server by a client. Clever implementation may be necessary.